requirements.txt
# Pinned for reproducibility
pyyaml>=6.0
librosa>=0.10.0
soundfile>=0.12.0
pandas>=1.5.0
scikit-learn>=1.2.0,<2.0
imbalanced-learn>=0.11.0
seaborn>=0.12.0
matplotlib>=3.6.0
joblib>=1.2.0
xgboost>=2.0.0
numpy>=1.23.0
pyaudio>=0.2.13
pytest>=7.0
